can add data to a DB and return to menu

+++ app/src/main/java/com/drinkjournal/dataClasses/DrinkData.kt
@@ -6,9 +6,9 @@ class DrinkData(
     var drinkName: String,
     var drinkType: String,
     var drinkSpecifics : String,
-    var drinkAlcoholPercentage: Short,
+    var drinkAlcoholPercentage: Int,
     var drinkMaker: String,
     var drinkOrigin: String,
     var drinkDescription: String,
-    var drinkRating: Int
+    var drinkRating: Float
 )
